+++ rgts-pricing-service/src/main/java/com/rahim/pricingservice/service/IAddGoldTypeService.java
@@ -6,6 +6,6 @@ import com.rahim.pricingservice.dto.request.AddGoldTypeRequest;
  * @author Rahim Ahmed
  * @created 23/03/2025
  */
-public interface IGoldTypeService {
+public interface IAddGoldTypeService {
     void addGoldType(AddGoldTypeRequest request);
 }
