+++ rgts-pricing-service/src/test/java/util/GoldCaratUtilTest.java
@@ -1,0 +1,8 @@
+package util;
+
+/**
+ * @author Rahim Ahmed
+ * @created 24/03/2025
+ */
+public class GoldCaratUtilTest {
+}
